+++ tools/initialize_es_document.py
@@ -14,17 +14,26 @@ os.environ.setdefault("DJANGO_CONFIGURATION", "Dev")
 # load AirOne application
 configurations.setup()
 
+from airone.lib.elasticsearch import ESS  # NOQA
 from entity.models import Entity  # NOQA
 from job.models import Job  # NOQA
 
 
-def update_es_document(entities):
+def initialize_es_document(entities):
+    es = ESS()
+
+    # clear previous index
+    es.indices.delete(index=es._index, ignore=[400, 404])
+
+    # create a new index with mapping
+    es.recreate_index()
+
     target_entity = Entity.objects.filter(is_active=True)
     if entities:
         target_entity = target_entity.filter(name__in=entities)
 
     for entity in target_entity:
-        Job.new_update_documents(entity).run()
+        Job.new_update_documents(entity, "", {"is_update": True}).run()
 
 
 def get_options():
@@ -36,4 +45,4 @@ def get_options():
 if __name__ == "__main__":
     (option, entities) = get_options()
 
-    update_es_document(entities)
+    initialize_es_document(entities)